pyproject.toml
[project]
name = "medical-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.2",
    "flask-limiter>=3.12",
    "langchain>=1.2.2",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.7",
    "langchain-pinecone>=0.2.13",
    "pypdf>=6.5.0",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.0",
    "cryptography>=44.0.0",
    "langchain-core>=1.2.7",
]
